+++ overlays/ranger.nix
@@ -14,9 +14,9 @@ self: super: {
 
     patches = (old.patches or [ ]) ++ [
       (super.fetchpatch {
-        name = "fix-ctrl-arrows-in-termite.patch";
-        url = "https://github.com/ranger/ranger/compare/master...Enzime:feature/support-modifier-arrow-keys.patch";
-        sha256 = "sha256-HongXSjIiVoD0l159Uu8sPhtc1iE0mpnGypjISFCc58=";
+        name = "fix-ctrl-arrows.patch";
+        url = "https://github.com/Enzime/ranger/commit/02d8c8f8500d46490d6724f92cf4e8ea89888d75.patch";
+        sha256 = "sha256-P48pn4vZgwW5JOhRm07a0/57+pNwpL+0JMDeyV9BRXg=";
       })
     ];
   });